% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amazon_reviews.R
\name{get_amazon_reviews}
\alias{get_amazon_reviews}
\title{Get Amazon Reviews}
\usage{
get_amazon_reviews(base_url, page_lim = 3, verbose = TRUE)
}
\arguments{
\item{base_url}{url provided by user}

\item{page_lim}{number of pages to parse}

\item{verbose}{whether to print current url}
}
\description{
Scrape Amazon reviews from a selected number of pages
}
