% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reviews.R
\name{get_reviews}
\alias{get_reviews}
\title{Get Reviews From Major Review Sites}
\usage{
get_reviews(base_url, page_lim = NULL, source, verbose = TRUE)
}
\arguments{
\item{base_url}{url provided by user}

\item{page_lim}{number of pages to parse}

\item{source}{the review source}

\item{verbose}{whether to print current url}
}
\description{
Scrape reviews from a selection of review sites.
}
\examples{

netflix <- "https://www.trustpilot.com/review/www.netflix.com"
get_reviews(base_url = netflix, page_lim = 2, source = "trustpilot")

snoop <- "https://www.amazon.com/Crook-Cook-Platinum-Recipes-Kitchen/product-reviews/1452179611/ref=cm_cr_dp_d_show_all_btm?ie=UTF8&reviewerType=all_reviews"
get_reviews(base_url = snoop, page_lim = 2, source = "amazon")

orsa <- "https://www.yelp.com/biz/osteria-dellorsa-bologna?osq=orsa"
get_reviews(base_url = orsa, page_lim = 2, source = "yelp")

}
