% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trustpilot_reviews.R
\name{get_trustpilot_reviews}
\alias{get_trustpilot_reviews}
\title{Get Trustpilot Reviews}
\usage{
get_trustpilot_reviews(base_url, page_lim = 3, company = NULL, verbose = TRUE)
}
\arguments{
\item{base_url}{url provided by user}

\item{page_lim}{number of pages to parse}

\item{company}{company name}

\item{verbose}{whether to print current url}
}
\description{
Scrape Trustpilot reviews from a selected number of pages
}
