% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_reviews}
\alias{get_reviews}
\title{Main function at user level}
\usage{
get_reviews(base_url, page_lim = NULL, source, verbose = TRUE)
}
\arguments{
\item{base_url}{url provided by user}

\item{page_lim}{number of pages to parse}

\item{source}{tre review source}

\item{verbose}{whether to print current url}
}
\description{
Main function at user level
}
\examples{
get_reviews("https://www.tripadvisor.co.uk/Attraction_Review-g189541-d245024-Reviews-The_Little_Mermaid_Den_Lille_Havfrue-Copenhagen_Zealand.html",2,source = "tripadvisor")
}
